+++ cmd/hw001/hw001.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+	helloWorldPrinter "github.com/rodkevich/go-course/homework/task001/hwp"
+)
+
+func main() {
+	var w = fmt.Println
+	helloWorldPrinter.AddEmoji(`:relaxed:`, w)
+}
