--- homework001/cmd/homework001.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	helloWorldPrinter "github.com/rodkevich/go-course/homework001/hwp"
-)
-
-func main() {
-	helloWorldPrinter.DisplayHelloWorldWithEmoji(`:relaxed:`)
-}
